+++ app/src/main/java/com/gammastudio/snowboardapp/Helmit.java
@@ -5,14 +5,14 @@ class Helmit {
     private String name;
     private String brand;
     private Float price;
-    private Integer rating;
+    private Float rating;
     private String imageURL;
 
     public Helmit(){
 
     }
 
-    public Helmit(Integer idVal, String nameVal, String brandVal, Float priceVal, Integer ratingVal, String imageURLVal){
+    public Helmit(Integer idVal, String nameVal, String brandVal, Float priceVal, Float ratingVal, String imageURLVal){
         this.id = idVal;
         this.name = nameVal;
         this.price = priceVal;
@@ -39,7 +39,7 @@ class Helmit {
     {
         return price;
     }
-    public Integer getRating()
+    public Float getRating()
     {
         return rating;
     }
@@ -65,7 +65,7 @@ class Helmit {
         this.price = price;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(Float rating) {
         this.rating = rating;
     }
